[Django] Fix: 100(10x10)초마다 데에터 전처리해서 경고 날려주는 SyncData 함수 수정

+++ Android/app/src/main/java/com/example/drinkingscanner/RetrofitService.java
@@ -18,7 +18,7 @@ public interface RetrofitService {
     Call<ServerResult> saveData(@FieldMap HashMap<String,Object> hm);
 
     @GET("preData")
-    Call<ServerResult> preData(@Query("user") String user, @Query("date") String date, @Query("bestSpeed") Integer bestSpeed);
+    Call<ServerResult> preData(@Query("user") String user, @Query("date") String date, @Query("bestSpeed") Float bestSpeed);
 
     @FormUrlEncoded
     @POST("syncData")
